Add FluencyBuilderTest#buildWithAllCustomConfig to fluency-td

+++ fluency-treasuredata/src/main/java/org/komamitsu/fluency/treasuredata/ingester/sender/TreasureDataSender.java
@@ -91,12 +91,12 @@ public class TreasureDataSender
         return client;
     }
 
-    public long getRetryInternalMs()
+    public int getRetryInternalMs()
     {
         return config.retryIntervalMs;
     }
 
-    public long getMaxRetryInternalMs()
+    public int getMaxRetryInternalMs()
     {
         return config.maxRetryIntervalMs;
     }
